requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
line-bot-sdk>=3.11.0,<4.0
anthropic>=0.40.0,<1.0
psycopg2-binary>=2.9.9,<3.0
requests>=2.31.0,<3.0
python-dotenv>=1.0.0,<2.0
google-api-python-client>=2.110.0,<3.0
google-auth>=2.27.0,<3.0
pypdf>=4.0.0,<6.0
apscheduler>=3.10.0,<4.0
